server/requirements.txt
certifi==2018.8.13
chardet==3.0.4
Django==2.1
idna==2.7
numpy==1.15.0
pandas==0.23.4
psycopg2==2.7.5
python-dateutil==2.7.3
python-scrapyd-api==2.1.2
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
